requirements.txt.txt
Flask
Flask-Cors
google-generativeai
beautifulsoup4
gunicorn
